move legend draw step to own module:
- un-expose legend.repositionText and legend.texts
- use anchor utils
